Changed: Incremented the project version by 0.1

+++ apps/acqdat_core/lib/acqdat_core/entity-management/schema/project.ex
@@ -24,7 +24,7 @@ defmodule AcqdatCore.Schema.EntityManagement.Project do
     field(:metadata, :map)
     field(:location, :map)
     field(:archived, :boolean, default: false)
-    field(:version, :integer, default: 1)
+    field(:version, :float, default: 1.0)
     field(:start_date, :utc_datetime)
 
     belongs_to(:org, Organisation, on_replace: :delete)
